+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Added
+- Direct support for handling browser based authentication in the AJAX spider.
 
 ## [0.8.0] - 2023-06-06
 ### Changed